playwright_service/pyproject.toml
[project]
name = "playwright-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi[standard]>=0.117.1",
    "lxml>=6.0.2",
    "pytest-playwright>=0.7.1",
]
